Extract permission screen into own Composable #24
