+++ app/src/main/java/app/zimly/backup/MainActivity.kt
@@ -1,16 +1,11 @@
 package app.zimly.backup
 
-import android.Manifest
 import android.content.pm.PackageManager
-import android.os.Build
 import android.os.Bundle
 import android.util.Log
 import androidx.activity.ComponentActivity
-import androidx.activity.compose.rememberLauncherForActivityResult
 import androidx.activity.compose.setContent
-import androidx.activity.result.contract.ActivityResultContracts
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.SideEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
@@ -25,6 +20,8 @@ import app.zimly.backup.data.remote.RemoteDao
 import app.zimly.backup.data.remote.ZimDatabase
 import app.zimly.backup.ui.screens.editor.EditorScreen
 import app.zimly.backup.ui.screens.list.ListScreen
+import app.zimly.backup.ui.screens.permission.PermissionScreen
+import app.zimly.backup.ui.screens.permission.getPermissions
 import app.zimly.backup.ui.screens.sync.SyncScreen
 import app.zimly.backup.ui.theme.ZimzyncTheme
 
@@ -70,29 +67,7 @@ class MainActivity : ComponentActivity() {
             // Grant permission for app
             // https://stackoverflow.com/questions/60608101/how-request-permissions-with-jetpack-compose
             composable(GRANT_PERMISSION) {
-                val permissionLauncher = rememberLauncherForActivityResult(
-                    ActivityResultContracts.RequestMultiplePermissions()
-                ) { isGranted ->
-
-                    // Lookup and compare the granted permission
-                    val granted = getPermissions()
-                        .map { permission -> isGranted[permission] }
-                        .reduce { granted, permission -> granted == true && permission == true }
-
-                    if (granted == true) {
-                        Log.i(localClassName, "Permissions granted")
-                        startDest = REMOTES_LIST
-                    } else {
-                        // TODO Implement some sort of informative screen, that the user
-                        // needs to grant permissions for the app to work.
-                        Log.i(localClassName, "PERMISSION DENIED")
-                    }
-                }
-                SideEffect {
-                    permissionLauncher.launch(
-                        getPermissions()
-                    )
-                }
+                PermissionScreen({startDest = REMOTES_LIST})
             }
 
             composable(REMOTES_LIST) {
@@ -159,25 +134,5 @@ class MainActivity : ComponentActivity() {
         return granted
     }
 
-    /*
-     * Returns the needed permissions based on Android SDK version.
-     * Changes were introduced to the media permissions in API 33+
-     */
-    private fun getPermissions(): Array<String> {
-        val permissions = if (Build.VERSION.SDK_INT < Build.VERSION_CODES.TIRAMISU) {
-            // Compatibility with older versions
-            arrayOf(
-                Manifest.permission.ACCESS_MEDIA_LOCATION,
-                Manifest.permission.READ_EXTERNAL_STORAGE
-            )
-        } else {
-            // New permissions
-            arrayOf(
-                Manifest.permission.READ_MEDIA_IMAGES,
-                Manifest.permission.READ_MEDIA_VIDEO,
-                Manifest.permission.ACCESS_MEDIA_LOCATION
-            )
-        }
-        return permissions
-    }
+
 }
